reuse component for show/new alert form and side panel alert settings
