+++ assets/js/components/alerts/AlertSetting.jsx
@@ -42,12 +42,12 @@ export default (props) => {
 
   const timeMenu = () => (
     <Menu onClick={e => { setValue(e.key) }}>
-      <Menu.Item key="15">15 mins</Menu.Item>
-      <Menu.Item key="30">30 mins</Menu.Item>
-      <Menu.Item key="60">1 hr</Menu.Item>
-      <Menu.Item key="360">6 hrs</Menu.Item>
-      <Menu.Item key="540">9 hrs</Menu.Item>
-      <Menu.Item key="1440">24 hrs</Menu.Item>
+      <Menu.Item key={15}>15 mins</Menu.Item>
+      <Menu.Item key={30}>30 mins</Menu.Item>
+      <Menu.Item key={60}>1 hr</Menu.Item>
+      <Menu.Item key={360}>6 hrs</Menu.Item>
+      <Menu.Item key={540}>9 hrs</Menu.Item>
+      <Menu.Item key={1440}>24 hrs</Menu.Item>
     </Menu>
   );
